+++ CHANGELOG.md
@@ -2,10 +2,14 @@
 
 ## v0.2.0
 
+### Added
+
+- Added a switch to generate the `spotdll.json` file from sample in the Current working directory. #3
+- Added a switch to define the JSON file to a completely seperate and unreleated path and file to the standard behaviour. #3
+
 ### Changed
 
 - The location of the file `spotdll.json` is now called from the folder the script is being called from, rather than from `~/bin`. #1
-- Added a switch to generate the `spotdll.json` file from sample in the Current working directory.
 
 ### Issues Closed