+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## v0.2.0
 
+### Added
+
+- Added a switch to generate the `spotdll.json` file from sample in the Current working directory. #3
+- Added a switch to define the JSON file to a completely seperate and unreleated path and file to the standard behaviour. #3
+- Added a switch to instruct the script to delete the JSON file after the task is complete. #3
+
 ### Changed
 
 - The location of the file `spotdll.json` is now called from the folder the script is being called from, rather than from `~/bin`. #1
@@ -9,6 +15,7 @@
 ### Issues Closed
 
 - #1
+- #3
 
 ---