+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added a switch to generate the `spotdll.json` file from sample in the Current working directory. #3
 - Added a switch to define the JSON file to a completely seperate and unreleated path and file to the standard behaviour. #3
 - Added a switch to instruct the script to delete the JSON file after the task is complete. #3
+- Added logging task. #4
 
 ### Changed
 
@@ -16,6 +17,7 @@
 
 - #1
 - #3
+- #4
 
 ---